move lib to parent folder
