+++ examples/jsm/nodes/materialx/lib/mx_hsv.js
@@ -1,4 +1,4 @@
-import { fn } from '../../../Nodes.js';
+import { fn } from '../../Nodes.js';
 
 // Original shader code from:
 // https://github.com/AcademySoftwareFoundation/MaterialX/blob/main/libraries/stdlib/genglsl/lib/mx_hsv.glsl
